Moved db interface to own file.
