+++ db/dbContext.go
@@ -12,16 +12,7 @@ import (
 	"github.com/go-pg/pg"
 )
 
-// DbContextInterface for DB access.
-type ContextInterface interface {
-	CreateSchema() error
-	GetModel(model models.Persistable) models.Persistable
-	GetWithWhere(model models.Persistable, refArray []interface{}, whereClause string, params ...interface{}) []interface{}
-	InsertModel(model models.Persistable) models.Persistable
-	UpdateModel(model models.Persistable) bool
-	DeleteModel(model models.Persistable) bool
-}
-
+// DbContext defines an abstract Datbase connection.
 type DbContext struct {
 	connection *pg.DB
 }
